子线程更新图片
深度图像传入CPP，处理后的图片更新在UI上
下一步纠正图片显示、处理问题

+++ NiViewer.Android/app/src/main/cpp/native-lib.cpp
@@ -27,7 +27,7 @@ Java_org_openni_android_tools_niviewer_StreamView_bitmap2Gray(JNIEnv *env, jobje
 
     // TODO
     //图像处理
-    Mat imgData(h,w,CV_8UC1,(unsigned char*)pixels);
+    Mat imgData(h,w,CV_16UC1,(unsigned char*)pixels);
     //cvtColor(imgData,imgData,CV_BGRA2GRAY);              //转灰度图
     cvtColor(imgData,imgData,CV_GRAY2BGRA);              //需要显示，再转回彩色图
     //applyColorMap(imgData,imgData,COLORMAP_HSV);
